Tagging release 1.2.5.

+++ test/autotest.py
@@ -34,11 +34,24 @@ RETRIES=2
 
 #Sleep after each program startup (sec)
 DEFAULT_S=0.3
+if sys.version_info[0] == 2 and sys.version_info[0:2] >= (2,7) and \
+    subprocess.check_output(['uname', '-p'])[0:3] == 'arm':
+  DEFAULT_S *= 2
+
 if testconfig.MTCP_USE_PROC_MAPS == "yes":
   DEFAULT_S = 2*DEFAULT_S
 S=DEFAULT_S
 #Appears as S*SLOW in code.  If --slow, then SLOW=5
 SLOW=1
+#In the case of gdb, even if both gdb and the inferior are running after
+#ckpt or restart, this does not guarantee that the ptrace-related work
+#(that is needed at resume or restart) is over. The ptrace related work happens
+#in the signal handler. Proceeding while still being inside the signal handler,
+#can lead to bad consquences. To play it on the safe side, PTRACE_SLEEP was
+#set at 2 seconds.  (Until this is fixed, --enable-ptrace-support will
+#remain experimental.)
+#if testconfig.PTRACE_SUPPORT == "yes":
+#  PTRACE_SLEEP=2
 
 #Max time to wait for ckpt/restart to finish (sec)
 TIMEOUT=10
@@ -252,6 +265,16 @@ os.system("rm -f "+tmpfile)
 
 os.environ['DMTCP_GZIP'] = GZIP
 
+# Temporary hack until DMTCP cleans up when using --enable-ptrace-support
+def deletePtraceFiles():
+  tmpdir = os.getenv("TMPDIR", "/tmp")  # if "TMPDIR" not set, return "/tmp"
+  tmpdir += "/dmtcp-" + pwd.getpwuid(os.getuid()).pw_name + \
+            "@" + socket.gethostname()
+  os.system("cd "+tmpdir+"; "+
+	    "rm -f ptrace_shared.txt ptrace_setoptions.txt \
+	     ptrace_ckpthreads.txt ptrace_shared.txt ptrace_setoptions.txt \
+	     ptrace_ckpthreads.txt new_ptrace_shared.txt ckpt_leader_file.txt")
+
 #launch the coordinator
 coordinator = launch(BIN+"dmtcp_coordinator")
 
@@ -462,6 +485,8 @@ def runTestRaw(name, numProcs, cmds):
       #  of this function:  testRestart
       testCheckpoint()
       printFixed("PASSED ")
+      #if testconfig.PTRACE_SUPPORT == "yes":
+      #  sleep(PTRACE_SLEEP)
       testKill()
 
       printFixed("rstr:")
@@ -469,6 +494,8 @@ def runTestRaw(name, numProcs, cmds):
         try:
           testRestart()
           printFixed("PASSED")
+          #if testconfig.PTRACE_SUPPORT == "yes":
+          #  sleep(PTRACE_SLEEP)
           break
         except CheckFailed, e:
           if j == RETRIES-1:
@@ -549,7 +576,13 @@ runTest("dmtcp1",        1, ["./test/dmtcp1"])
 
 runTest("dmtcp2",        1, ["./test/dmtcp2"])
 
+# dmtcp3 creates 10 threads; Keep checkpoint image small by using gzip
+# Also, it needs some extra time to startup
+S=2
+os.environ['DMTCP_GZIP'] = "1"
 runTest("dmtcp3",        1, ["./test/dmtcp3"])
+os.environ['DMTCP_GZIP'] = GZIP
+S=DEFAULT_S
 
 runTest("dmtcp4",        1, ["./test/dmtcp4"])
 
@@ -614,9 +647,9 @@ os.environ['DMTCP_GZIP'] = "0"
 runTest("shared-memory", 2, ["./test/shared-memory"])
 
 # This is arguably a bug in the Linux kernel 3.2 for ARM.
-if sys.version_info[0] == 2 and sys.version_info[0:2] >= (2,7):
-  if subprocess.check_output(['uname', '-p'])[0:3] == 'arm':
-    print "On ARM, there is a known issue with the sysv-shm test."
+if sys.version_info[0] == 2 and sys.version_info[0:2] >= (2,7) and \
+    subprocess.check_output(['uname', '-p'])[0:3] == 'arm':
+  print "On ARM, there is a known issue with the sysv-shm test. Not running it."
 runTest("sysv-shm",      2, ["./test/sysv-shm"])
 
 #Invoke this test when we drain/restore data in pty at checkpoint time.
@@ -635,7 +668,7 @@ runTest("pthread2",      1, ["./test/pthread2"])
 S=3
 runTest("pthread3",      1, ["./test/pthread2 80"])
 S=DEFAULT_S
-runTest("pthread4",      1, ["./test/pthread4"])
+runTest("pthread4",      1, ["./test/pthread4 80"])
 
 os.environ['DMTCP_GZIP'] = "1"
 runTest("gzip",          1, ["./test/dmtcp1"])
@@ -679,6 +712,8 @@ if testconfig.HAS_VIM == "yes" and testconfig.PID_VIRTUALIZATION == "yes":
     # Delete previous vim processes.  Vim behaves poorly with stale processes.
     vimCommand = testconfig.VIM + " /etc/passwd +3" # +3 makes cmd line unique
     def killCommand(cmdToKill):
+      if os.getenv('USER') == None:
+        return
       ps = subprocess.Popen(['ps', '-u', os.environ['USER'], '-o', 'pid,command'],
     		            stdout=subprocess.PIPE).communicate()[0]
       for row in ps.split('\n')[1:]:
@@ -724,25 +759,32 @@ if testconfig.HAS_SCREEN == "yes" and testconfig.PID_VIRTUALIZATION == "yes":
                                 " -c /dev/null -s /bin/sh"])
   S=DEFAULT_S
 
-if testconfig.PTRACE_SUPPORT == "yes" and sys.version_info[0:2] >= (2,6):
-  if testconfig.HAS_STRACE == "yes":
+if testconfig.PTRACE_SUPPORT == "yes" and \
+   (testconfig.HAS_STRACE == "yes" or testconfig.HAS_GDB == "yes"):
+  print "  Deleting files in /tmp/dmtcp-USER@HOST before ptrace tests.  (Until"
+  print "  this is fixed, --enable-ptrace-support will remain experimental.)"
+  deletePtraceFiles()
+  if testconfig.HAS_STRACE == "yes" and testconfig.PTRACE_SUPPORT == "yes":
     S=1
-    runTest("strace",    2,  ["strace test/dmtcp2"])
+    if sys.version_info[0:2] >= (2,6):
+      runTest("strace",    2,  ["strace test/dmtcp2"])
     S=DEFAULT_S
 
-  if testconfig.HAS_GDB == "yes":
+  deletePtraceFiles()
+  if testconfig.HAS_GDB == "yes" and testconfig.PTRACE_SUPPORT == "yes":
     os.system("echo 'run' > dmtcp-gdbinit.tmp")
     S=2
-    runTest("gdb",          2, ["gdb -n -batch -x dmtcp-gdbinit.tmp test/dmtcp1"])
+    if sys.version_info[0:2] >= (2,6):
+      runTest("gdb",       2,  ["gdb -n -batch -x dmtcp-gdbinit.tmp test/dmtcp1"])
+    S=DEFAULT_S
+    os.system("rm -f dmtcp-gdbinit.tmp")
 
-    runTest("gdb-pthread0", 2, ["gdb -n -batch -x dmtcp-gdbinit.tmp test/dmtcp3"])
-
-    # These tests currently fail sometimes (if the computation is checkpointed
-    # while a thread is being created). Re-enable them when this issue has been
-    # fixed in the ptrace module.
-    #runTest("gdb-pthread1", 2, ["gdb -n -batch -x dmtcp-gdbinit.tmp test/pthread1"])
-    #runTest("gdb-pthread2",2, ["gdb -n -batch -x dmtcp-gdbinit.tmp test/pthread2"])
-
+  deletePtraceFiles()
+  if testconfig.HAS_GDB == "yes" and testconfig.PTRACE_SUPPORT == "yes":
+    os.system("echo 'run' > dmtcp-gdbinit.tmp")
+    S=2
+    if sys.version_info[0:2] >= (2,6):
+      runTest("gdb-pthread",2,  ["gdb -n -batch -x dmtcp-gdbinit.tmp test/dmtcp3"])
     S=DEFAULT_S
     os.system("rm -f dmtcp-gdbinit.tmp")
